Legacy model updates, started on current model
Worked through most of the SmileMod1.bas file and Smile.frm to see how some of the basic loading works. This lead to creating the legacy model and the current model to start some of the data migration. Also, downloaded the SMILE RF guide html. This is a local user's guide which is VERY HELPFUL for learning about memory layout and the meaning of PLM's, BTS, and the majority of the UI displays were straight forward - so this should be the best guide to figuring out the legacy source code.